+++ src/main/java/com/lipcam/PedidosApiSpring/repositories/ProdutosRepository.java
@@ -3,6 +3,6 @@ package com.lipcam.PedidosApiSpring.repositories;
 import com.lipcam.PedidosApiSpring.entities.Produtos;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProdutosRepository extends JpaRepository<Produtos, Integer> {
+public interface ProdutosRepository extends JpaRepository<Produtos, Long> {
     Produtos findByDescricao(String descricao);
 }
